Moved arch() into BuildOutput.
